+++ packages/manager/src/features/Profile/DisplaySettings/DisplaySettings.tsx
@@ -1,130 +1,129 @@
-import { Profile } from '@linode/api-v4/lib/profile';
-import { path, pathOr } from 'ramda';
+import { updateUser } from '@linode/api-v4/lib/account';
 import * as React from 'react';
-import { connect, MapDispatchToProps } from 'react-redux';
-import { compose as recompose } from 'recompose';
-import {
-  createStyles,
-  Theme,
-  withStyles,
-  WithStyles
-} from 'src/components/core/styles';
-import setDocs from 'src/components/DocsSidebar/setDocs';
-import { DocumentTitleSegment } from 'src/components/DocumentTitle';
-import { AccountsAndPasswords } from 'src/documentation';
-import { updateProfile as handleUpdateProfile } from 'src/store/profile/profile.requests';
-import { MapState } from 'src/store/types';
-import EmailChangeForm from './UsernameAndEmailForm';
+import { useSelector } from 'react-redux';
+import { useLocation } from 'react-router-dom';
+import { compose } from 'recompose';
+import Paper from 'src/components/core/Paper';
+import { makeStyles, Theme } from 'src/components/core/styles';
+import { SingleTextFieldForm } from 'src/components/SingleTextFieldForm/SingleTextFieldForm';
+import useAccountManagement from 'src/hooks/useAccountManagement';
+import useProfile from 'src/hooks/useProfile';
+import { ApplicationState } from 'src/store';
+import withNotifications, {
+  WithNotifications
+} from 'src/store/notification/notification.containers';
+import useTimezone from 'src/utilities/useTimezone';
+import { v4 } from 'uuid';
 import TimezoneForm from './TimezoneForm';
 
-type ClassNames = 'root' | 'title';
+const useStyles = makeStyles((theme: Theme) => ({
+  paper: {
+    padding: theme.spacing(3),
+    paddingBottom: theme.spacing(3),
+    marginBottom: theme.spacing(3)
+  }
+}));
 
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {
-      padding: theme.spacing(3),
-      paddingBottom: theme.spacing(3)
-    },
-    title: {
-      marginBottom: theme.spacing(2)
+export const DisplaySettings: React.FC<WithNotifications> = props => {
+  const classes = useStyles();
+
+  const { updateProfile, requestProfile } = useProfile();
+  const { profile, _isRestrictedUser } = useAccountManagement();
+  const timezone = useTimezone();
+  const loggedInAsCustomer = useSelector(
+    (state: ApplicationState) => state.authentication.loggedInAsCustomer
+  );
+  const location = useLocation();
+
+  const emailRef = React.createRef<HTMLInputElement>();
+
+  React.useEffect(() => {
+    if (location.state?.focusEmail && emailRef.current) {
+      emailRef.current.focus();
+      emailRef.current.scrollIntoView();
     }
-  });
+  }, [emailRef, location.state]);
 
-interface State {
-  submitting: boolean;
-  updatedEmail: string;
-  errors?: any;
-  success?: any;
-}
+  // Used as React keys to force-rerender forms.
+  const [emailResetToken, setEmailResetToken] = React.useState(v4());
+  const [usernameResetToken, setUsernameResetToken] = React.useState(v4());
+  const [timezoneResetToken, setTimezoneResetToken] = React.useState(v4());
 
-type CombinedProps = StateProps & DispatchProps & WithStyles<ClassNames>;
-
-export class DisplaySettings extends React.Component<CombinedProps, State> {
-  state: State = {
-    submitting: false,
-    updatedEmail: this.props.email || '',
-    errors: undefined,
-    success: undefined
-  };
-
-  render() {
-    const {
-      email,
-      updateProfile,
-      loggedInAsCustomer,
-      loading,
-      timezone,
-      username
-    } = this.props;
-
-    if (!email || !username) {
-      return null;
-    }
-
-    return (
-      <div>
-        <DocumentTitleSegment segment="Display" />
-        {!loading && (
-          <React.Fragment>
-            <EmailChangeForm
-              email={email}
-              username={username}
-              updateProfile={updateProfile}
-              data-qa-email-change
-            />
-            <TimezoneForm
-              timezone={timezone}
-              updateProfile={updateProfile}
-              loggedInAsCustomer={loggedInAsCustomer}
-            />
-          </React.Fragment>
-        )}
-      </div>
-    );
+  if (!profile.data?.username) {
+    return null;
   }
 
-  static docs = [AccountsAndPasswords];
-}
-
-const styled = withStyles(styles);
-
-interface StateProps {
-  loading: boolean;
-  username?: string;
-  email?: string;
-  timezone: string;
-  loggedInAsCustomer: boolean;
-}
-
-const mapStateToProps: MapState<StateProps, {}> = state => {
-  const { profile } = state.__resources;
-  return {
-    loading: profile.loading,
-    username: path(['data', 'username'], profile),
-    email: path(['data', 'email'], profile),
-    timezone: pathOr<string>('GMT', ['data', 'timezone'], profile),
-    loggedInAsCustomer: pathOr(
-      false,
-      ['authentication', 'loggedInAsCustomer'],
-      state
-    )
+  const updateUsername = (newUsername: string) => {
+    setEmailResetToken(v4());
+    setTimezoneResetToken(v4());
+    return updateUser(profile?.data?.username ?? '', {
+      username: newUsername
+    });
   };
+
+  const updateEmail = (newEmail: string) => {
+    setUsernameResetToken(v4());
+    setTimezoneResetToken(v4());
+    return updateProfile({ email: newEmail });
+  };
+
+  const updateTimezone = (newTimezone: string) => {
+    setUsernameResetToken(v4());
+    setEmailResetToken(v4());
+    return updateProfile({ timezone: newTimezone });
+  };
+
+  return (
+    <>
+      <Paper className={classes.paper}>
+        <SingleTextFieldForm
+          key={usernameResetToken}
+          label="Username"
+          submitForm={updateUsername}
+          initialValue={profile.data.username}
+          disabled={_isRestrictedUser}
+          tooltipText={
+            _isRestrictedUser
+              ? 'Restricted users cannot update their username. Please contact an account administrator.'
+              : undefined
+          }
+          cb={requestProfile}
+        />
+      </Paper>
+
+      <Paper className={classes.paper}>
+        <SingleTextFieldForm
+          key={emailResetToken}
+          label="Email"
+          submitForm={updateEmail}
+          initialValue={profile.data.email}
+          cb={() => {
+            // If there's a "user_email_bounce" notification for this user, and
+            // the user has just updated their email, re-request notifications to
+            // potentially clear the email bounce notification.
+            const hasUserEmailBounceNotification = props.notifications.find(
+              thisNotification => thisNotification.type === 'user_email_bounce'
+            );
+            if (hasUserEmailBounceNotification) {
+              props.requestNotifications();
+            }
+          }}
+          inputRef={emailRef}
+          type="email"
+        />
+      </Paper>
+      <Paper className={classes.paper}>
+        <TimezoneForm
+          key={timezoneResetToken}
+          timezone={timezone}
+          loggedInAsCustomer={loggedInAsCustomer}
+          updateTimezone={updateTimezone}
+        />
+      </Paper>
+    </>
+  );
 };
 
-interface DispatchProps {
-  updateProfile: (v: Profile) => Promise<Profile>;
-}
-
-const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = dispatch => ({
-  updateProfile: (v: Profile) => dispatch(handleUpdateProfile(v) as any)
-});
-
-const connected = connect(mapStateToProps, mapDispatchToProps);
-
-const enhanced = recompose<CombinedProps, {}>(
-  styled,
-  connected,
-  setDocs(DisplaySettings.docs)
-);
+const enhanced = compose<WithNotifications, {}>(withNotifications());
 
 export default enhanced(DisplaySettings);
